Port to .NET 10, fix 20 defects, and finish the unimplemented targets
The tool was built on netcoreapp3.1, out of support since 2022, and had no way
to tell a change in output from a break in it. That came first: a fixture
generator, golden trees for every artifact, and toolchain gates that compile
and run what the generators emit. Everything after leans on those.

Port
  netcoreapp3.1 -> net10.0. Packages moved forward and a transitive
  System.Security.Cryptography.Xml 8.0.2 with seven high-severity advisories
  was lifted to a patched version. CI installed an EOL .NET 5 and its test
  step was a no-op; the build scripts passed -r with no value and one of them
  published Debug. All rewritten.

Defects
  Twenty, of which six were found by the new gates rather than by reading.
  Three were the same shape - a value corrupted so quietly that the obvious
  check agreed with it:
    - the binary writer cast every 64-bit value through uint, truncating it
    - JSON exported int64 as a number, which JSON.parse rounds
    - the JSON path discarded a float's 32-bit precision
  The first only surfaced once a second implementation existed to read what
  the first had written. Also: the documented foreign RefTable.RefFieldName
  form never worked, ragged sheets crashed the importer, Excel date cells
  arrived as serial numbers, parsing depended on the machine's locale, cell
  references past column X were misreported, and generated TypeScript did not
  compile whenever a sheet declared a const entity or a record reference.

Features that had been left unimplemented
  TargetSide now filters output rather than only appearing in the docs.
  Validation runs at all - it was a 204-line method nothing called, whose
  uniqueness check skipped exactly the fields it meant to inspect - and
  reports every problem in one pass instead of one per run.
  Delimited array types, alongside the existing fixed-width serial fields.
  C++ and TypeScript binary readers, and generators for both.
  MySQL, PostgreSQL, MongoDB and Redis exporters that load directly, each
  swapping shadow storage in atomically. Connection secrets come from the
  environment so a recipe can be committed.

Structure
  The exporter carried its own writer and the generated C# its own reader,
  instead of both borrowing a 3,600-line runtime that a Unity project had to
  install as a plugin and of which the generated code called four members.
  Every target's output is now self-contained. Both are Span-based and free of
  per-value allocations.

Tests
  78, against real toolchains and real database engines. The suite writes with
  one implementation and reads with another, which is the only thing that
  catches the readers drifting from the writer.

+++ lib/Unity/SheetManForUnity/Assets/Plugins/SheetMan.Runtime/LiteBinaryExtensions.cs
@@ -101,7 +101,12 @@ namespace SheetMan.Runtime
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static void Write(this LiteBinaryWriter writer, long value)
         {
-            writer.WriteFixed64((uint)value);
+            // (ulong), not (uint). Casting through uint truncated every 64-bit value
+            // to its low 32 bits before widening again, so anything outside
+            // [0, uint.MaxValue] was silently corrupted on write - negative values
+            // came back as large positives, and large positives lost their high half.
+            // The reader was always reading a full Fixed64, so only the writer was wrong.
+            writer.WriteFixed64((ulong)value);
         }
 
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
